game by Accelerometer sensor and Proximity sensor
The game consists of two interfaces.
The first interface contains a box for entering questions, a button to add a question, a text displaying the number of questions entered, and a button to start playing.
When you press the Add Question button and the box is empty, a message must be shown to the user (type a question in the top box).
When the start playing button is pressed and the number of questions is less than 2, a message should be shown to the user (two or more questions must be added).
When you press the start playing button and the number of questions is 2 or more, you will go to the second interface. 
The second interface that contains the display space sequentially and continuously.
When you put the finger on the questions to repeat again.
It ends and the app returns to the game interface directly when the device screen is directed downwards (towards the negative Z axis).